Move `Count` struct to separate file.
